requirements.txt
h5py>=2.10.0,
jupyter>=1.0.0,
matplotlib>=3.2.1,
nas-bench-201==2.0,
numpy>=1.18.4,
prettytable>=2.0.0,
pytorch-ignite>=0.3.0,
pytorchcv>=0.0.58,
scikit-learn>=0.23.2,
scipy>=1.4.1,
tqdm>=4.46.0,
seaborn